+++ CHANGELOG.md
@@ -10,10 +10,14 @@ Unreleased
  * [#134] `Tail` plugin
  * [#135] Allow using the built-in plugins without the `plugin` feature.
  * [#137] Allow using multiple plugins.
+ * [#136] `Cursor` is now public.
+ * [#136] Added `TextBox::take_plugins()`.
 
 ## Changed:
 
  * **breaking** [#133] `TextBoxStyle` and `TextBoxStyleBuilder` no longer implement the `Default` trait.
+ * [#136] Replaced `TextBoxProperties::box_height` with  in `TextBoxProperties::bounding_box`.
+ * [#136] Reworked the editor example to support vertical cursor movement and mouse input.
 
 ## Fixed:
 
@@ -26,6 +30,7 @@ Unreleased
 [#133]: https://github.com/embedded-graphics/embedded-text/pull/133
 [#134]: https://github.com/embedded-graphics/embedded-text/pull/134
 [#135]: https://github.com/embedded-graphics/embedded-text/pull/135
+[#136]: https://github.com/embedded-graphics/embedded-text/pull/136
 [#137]: https://github.com/embedded-graphics/embedded-text/pull/137
 [#140]: https://github.com/embedded-graphics/embedded-text/pull/140